--- a/justfile
+++ b/justfile
@@ -19,7 +19,7 @@ docs:
 
 # Execute all unit tests in the workspace
 test:
-    cargo llvm-cov nextest
+    cargo llvm-cov nextest --features="parallel"
 
 # Run the entire CI pipeline including format, clippy, docs, and test checks
 ci: format clippy docs test