profiler:  https://share.firefox.dev/41Yjhbd 

```
samply record {path of target of test}
```

## Description

- adjust parallel iteration using `rayon` when logic in loop is complex (calling C++ / complex more than just insert element on array)
   - if we use par iter too much dept, from context switching using parallel execution, there is risk on cpu might halt. So tried not to use par_iter on not necessary loop. 
   - also for same reason, instead using `nextest` which spawn process per test, revert back to `cargo test` for CI
- minor improvement on memory including 
   - `&T` * `&T` implementation to avoid cloning 
   - avoid non necessary `const_zero`(which calls C++) on Poly or PolyMartix operation

## Perf

### before
```
Peak memory consumptions 542539776  bytes (0.51 GB)
Time to obfuscate: 105.465769375s
Time for evaluation: 116.751426959s
Total time: 220.925412584s
```

### after

command with `parallel`
```
 /usr/bin/time -l cargo test test_io_just_mul_enc_and_bit --release --package diamond-io --lib --features="parallel" -- --nocapture
```
result
```
Peak memory consumptions 30360640  bytes (0.03036064 GB)
Time to obfuscate: 42.85817575s
Time for evaluation: 19.117325s
Total time: 61.97550075s
```

command without `parallel`
```
 /usr/bin/time -l cargo test test_io_just_mul_enc_and_bit --release --package diamond-io --lib -- --nocapture
 ```
 
 result
 ```
Peak memory consumptions 30049664  bytes (0.03 GB)
 Time to obfuscate: 103.073604208s
Time for evaluation: 80.958975542s
Total time: 184.03257975s
 ```